feat: Add CartPage component and routing
