+++ mission/src/router/index.js
@@ -1,27 +1,5 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import ItemListPage from '@/views/ItemList.vue';
-import ItemInfoPage from '@/views/ItemInfo.vue';
-
-const routes = [
-  {
-    path: '/',
-    name: 'Home',
-    component: ItemListPage,
-  },
-  {
-    path: '/about',
-    name: 'About',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/About.vue'),
-  },
-  {
-    path: '/item/:id',
-    name: 'ItemPage',
-    component: ItemInfoPage,
-  },
-];
+import routes from './routes';
 
 const router = createRouter({
   history: createWebHistory(process.env.BASE_URL),
